tested MessageActionMapper
